+++ nbproject/private/private.xml
@@ -5,12 +5,12 @@
             <url>src/proyecto_gm/Empleado/DatosEmpleados.java</url>
             <bookmark id="3">
                 <name/>
-                <line>211</line>
+                <line>274</line>
                 <key/>
             </bookmark>
             <bookmark id="5">
                 <name/>
-                <line>258</line>
+                <line>419</line>
                 <key/>
             </bookmark>
         </file>
@@ -41,39 +41,35 @@
     </editor-bookmarks>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-<<<<<<< HEAD
-            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/RecibosHonorarios/DatosRecibosHonorarios.java</file>
-            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/RecibosHonorarios/frmListaRecibosHonorarios.java</file>
-            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/Articulo/frmArticulo.java</file>
-            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/ConexionBD.java</file>
-            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/Asistencias/frmAsistenciaImagen.java</file>
-            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/RecibosHonorarios/ReciboHonorario.java</file>
-            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/RecibosHonorarios/frmRecibosHonorarios.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/Contactos/frmListaContacto.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/seguridad/frmGestionBD.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/Departamentos/DatosDepartamentos.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/Contactos/Contacto.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/seguridad/frmUsuario.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/Contactos/DatosContacto.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/Cargo/frmCargo.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/seguridad/frmRespaldarBD.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/Contactos/frmContactos.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/inicio.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/Cargo/DatosCargo.java</file>
         </group>
         <group>
-            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/RecibosHonorarios/DatosRecibosHonorarios.java</file>
-            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/RecibosHonorarios/frmListaRecibosHonorarios.java</file>
-            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/ConexionBD.java</file>
-            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/RecibosHonorarios/ReciboHonorario.java</file>
-            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/RecibosHonorarios/frmRecibosHonorarios.java</file>
-=======
-            <file>file:/D:/trabajo/proyecto_gm/src/proyecto_gm/Utilitario.java</file>
-            <file>file:/D:/trabajo/proyecto_gm/src/proyecto_gm/seguridad/frmGestionBD.java</file>
-            <file>file:/D:/trabajo/proyecto_gm/src/proyecto_gm/Cargo/frmCargo.java</file>
-            <file>file:/D:/trabajo/proyecto_gm/src/proyecto_gm/Carreras/frmCarreras.java</file>
-            <file>file:/D:/trabajo/proyecto_gm/src/proyecto_gm/Empleado/DatosEmpleados.java</file>
-            <file>file:/D:/trabajo/proyecto_gm/src/proyecto_gm/ConexionBD.java</file>
-            <file>file:/D:/trabajo/proyecto_gm/src/proyecto_gm/inicio.java</file>
-            <file>file:/D:/trabajo/proyecto_gm/src/proyecto_gm/seguridad/frmRespaldarBD.java</file>
-            <file>file:/D:/trabajo/proyecto_gm/src/proyecto_gm/seguridad/frmUsuario.java</file>
-            <file>file:/D:/trabajo/proyecto_gm/src/proyecto_gm/Area/frmArea.java</file>
-            <file>file:/D:/trabajo/proyecto_gm/src/proyecto_gm/Empleado/frmListaEmpleado.java</file>
-            <file>file:/D:/trabajo/proyecto_gm/src/proyecto_gm/Empleado/Empleados.java</file>
-            <file>file:/D:/trabajo/proyecto_gm/src/proyecto_gm/Empleado/frmEmpleado.java</file>
-        </group>
-        <group>
-            <file>file:/D:/trabajo/proyecto_gm/src/proyecto_gm/ConexionBD.java</file>
->>>>>>> ab228c78dd13e54d26baa21e600be29cdab9bcb2
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/Contactos/frmListaContacto.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/Proveedores/DatosProveedores.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/Cargo/frmCargo.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/Instituciones/frmInstituciones.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/Proveedores/frmProveedores.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/Contactos/frmContactos.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/Proveedores/frmListaProveedores.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/Cargo/DatosCargo.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/seguridad/frmGestionBD.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/Instituciones/frmListaInstituciones.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/Departamentos/DatosDepartamentos.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/Contactos/Contacto.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/seguridad/frmUsuario.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/Contactos/DatosContacto.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/seguridad/frmRespaldarBD.java</file>
+            <file>file:/C:/Users/elasp/Downloads/Prueba_Tecnica/proyecto_gm/src/proyecto_gm/inicio.java</file>
         </group>
     </open-files>
 </project-private>
